# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "register new user",
                "description": "register new user",
                "operationId": "registerNewUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true
                    },
                    {
                        "name": "role_name",
                        "in": "path",
                        "description": "role name",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "path",
                        "description": "first name (for client)",
                        "required": true
                    },
                    {
                        "name": "last_name",
                        "in": "path",
                        "description": "last name (for client)",
                        "required": true
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "description": "username (for client)",
                        "required": true
                    },
                    {
                        "name": "phone_number",
                        "in": "path",
                        "description": "phone number (for client)(for organization)",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name (for organization)",
                        "required": true
                    },
                    {
                        "name": "postal_code",
                        "in": "path",
                        "description": "postal code (for organization)",
                        "required": true
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "postal code (for organization)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "log in user",
                "description": "log in user",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "name": "emailorusername",
                        "in": "path",
                        "description": "email or username",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "log out user",
                "description": "log out user",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get authenticated user info",
                "description": "get authenticated user info",
                "operationId": "UserInfo",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}